requirements.txt
asgiref==3.3.1
certifi==2020.11.8
Django==3.1.3
django-autofixture==0.12.1
djangorestframework==3.11.0
pytz==2020.4
sqlparse==0.4.1
six==1.15.0
psycopg2-binary>=2.8
python-dotenv==0.13.0
drf-yasg==1.17.1